Renamed command-functions.js -> commands.js
